+++ service/zpages.go
@@ -40,7 +40,7 @@ func (srv *service) RegisterZPages(mux *http.ServeMux, pathPrefix string) {
 }
 
 func (srv *service) handleServicezRequest(w http.ResponseWriter, r *http.Request) {
-	r.ParseForm()
+	r.ParseForm() // nolint:errcheck
 	w.Header().Set("Content-Type", "text/html; charset=utf-8")
 	zpages.WriteHTMLHeader(w, zpages.HeaderData{Title: "service"})
 	zpages.WriteHTMLComponentHeader(w, zpages.ComponentHeaderData{
@@ -58,7 +58,7 @@ func (srv *service) handleServicezRequest(w http.ResponseWriter, r *http.Request
 }
 
 func (srv *service) handlePipelinezRequest(w http.ResponseWriter, r *http.Request) {
-	r.ParseForm()
+	r.ParseForm() // nolint:errcheck
 	w.Header().Set("Content-Type", "text/html; charset=utf-8")
 	pipelineName := r.Form.Get(zPipelineName)
 	componentName := r.Form.Get(zComponentName)
@@ -116,7 +116,7 @@ func (srv *service) getPipelinesSummaryTableData() zpages.SummaryPipelinesTableD
 }
 
 func handleExtensionzRequest(host component.Host, w http.ResponseWriter, r *http.Request) {
-	r.ParseForm()
+	r.ParseForm() // nolint:errcheck
 	w.Header().Set("Content-Type", "text/html; charset=utf-8")
 	extensionName := r.Form.Get(zExtensionName)
 	zpages.WriteHTMLHeader(w, zpages.HeaderData{Title: "Extensions"})